edge/requirements.txt
fastapi>=0.109.0,<1.0.0
uvicorn[standard]>=0.25.0,<1.0.0
pydantic>=2.6.0,<3.0.0
pydantic-settings>=2.1.0,<3.0.0
numpy>=1.26.0,<2.0.0
Pillow>=10.2.0,<11.0.0
